requirements.txt
requests
fastapi
uvicorn
python-decouple